The other players get their own hook: search, invite, challenge, kick, nudge
One search box serves two jobs (#311), so the filter, the two sends that clear
it, and the roster writes that follow travel together — with the two
viewer-keyed reads the filter depends on. It reads useAuth() itself rather than
taking the viewer as a parameter, which is what keeps its three effects inside
the #1390 ratchet; the ratchet's registry gains the new file.
